Django/Django_4/pyproject.toml
[tool.poetry]
name = "django_project"
version = "0.1.0"
description = "python django project"
authors = ["sub-blind <rlawotjq206@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
